PDCurseControls - all moved out of separate include/src folders and into same folder. All projects were updated with change Added BattleState to main project. This required updating the BattleProcessor to have a default constructor. GameState now gets a pointer to the GameStateManager.

+++ RPGLib/DataPkg.cpp
@@ -141,12 +141,12 @@ void DataPkg::blockOp(bool op, const void* datum, int datumOffset, int size, int
 
 void DataPkg::appendData(const void* datum, int size)
 {
-	writeData(datum, size, appendOffset);
+	writeData(datum, size, (int)appendOffset);
 }
 
 void DataPkg::readNextData(const void* datum, int size)
 {
-	readData(datum, size, readOffset);
+	readData(datum, size, (int)readOffset);
 }
